Refactor file upload functionality: - Removed legacy upload.ts file and migrated its logic to a new S3-based implementation. - Introduced new configuration and controller files for S3 integration. - Implemented file upload service with support for multiple input types (network, local, base64, buffer). - Added validation schemas using Zod for input and output. - Created worker support for file service initialization and upload handling. - Enhanced logging for better traceability during file operations.
